🔨 refactor: 답장 보내기 API 테스트 분리
