+++ src/pages/LetterReceptionPage/NormalReception/ReplyToLetter/index.test.tsx
@@ -4,15 +4,12 @@ import { HttpResponse, http } from 'msw';
 import { zodResolver } from '@hookform/resolvers/zod';
 import { server } from '@/mocks/node';
 import { baseURL } from '@/utils/mswUtils';
-import { API_PATHS, ROUTER_PATHS } from '@/constants/routerPaths';
+import { API_PATHS } from '@/constants/routerPaths';
 import { ReceptionLetter as ReceptionLetterData } from '@/mocks/datas/letter';
-import { render, screen, waitFor, renderHook } from '@/utils/testing-library';
+import { render, screen } from '@/utils/testing-library';
 import { formatDate } from '@/utils/dateUtils';
 import { WORRY_DICT } from '@/constants/users';
-import useBoolean from '@/hooks/useBoolean';
-import { letterResolvers } from '@/mocks/resolvers/letter';
 import { ReplyInputs, replySchema } from '..';
-import ReplyBottomSheet from './ReplyBottomSheet';
 import ReplyToLetter from '.';
 
 const ResizeObserver = vi.fn(() => ({
@@ -188,54 +185,5 @@ describe('편지 작성 테스트', () => {
 
     const bottomSheet = screen.getByText('답장을 보낼까요?');
     expect(bottomSheet).toBeVisible();
-  });
-});
-
-describe('답장 보내기 API 테스트', () => {
-  const letter_id = 2;
-  const { result: bottomSheetProps } = renderHook(() => useBoolean(true));
-
-  const ReplyToLetterComponent = () => {
-    const methods = useForm<ReplyInputs>({
-      resolver: zodResolver(replySchema),
-      defaultValues: {
-        replyContent: '답장을 보낼 편지 내용입니다.',
-      },
-    });
-    return (
-      <FormProvider {...methods}>
-        <ReplyBottomSheet {...bottomSheetProps.current} letterId={letter_id} />
-      </FormProvider>
-    );
-  };
-
-  it('[성공] 답장 보내기 버튼을 클릭하면 API 호출이 발생하고, 요청이 성공하면 루트 페이지로 이동하고 성공 토스트가 나타난다.', async () => {
-    const { user } = render(<ReplyToLetterComponent />);
-
-    const sendButton = screen.getByRole('button', { name: '보내기' });
-    await user.click(sendButton);
-
-    await waitFor(() => {
-      expect(navigateFn).toHaveBeenCalledWith(ROUTER_PATHS.ROOT);
-      expect(toast.success).toHaveBeenCalledTimes(1);
-    });
-  });
-  it('[실패] 답장 보내기 버튼을 클릭하면 API 호출이 발생하고, 요청이 실패하면 루트 페이지로 이동하고 실패 토스트가 나타난다.', async () => {
-    server.use(
-      http.patch(
-        baseURL(API_PATHS.LETTER_RECEPTION_REPLY_DETAIL(letter_id.toString())),
-        letterResolvers.patchLetterReceptionReplyDetail.alreadyReplyExist,
-      ),
-    );
-
-    const { user } = render(<ReplyToLetterComponent />);
-
-    const sendButton = screen.getByRole('button', { name: '보내기' });
-    await user.click(sendButton);
-
-    await waitFor(() => {
-      expect(navigateFn).toHaveBeenCalledWith(ROUTER_PATHS.ROOT);
-      expect(toast.error).toHaveBeenCalledTimes(1);
-    });
   });
 });
